# openapi.yaml
{
  "openapi": "3.0.0",
  "info": {
    "title": "Spark ETL API",
    "version": "1.0.0",
    "description": "API documentation for the Spark ETL application, generated from server actions."
  },
  "servers": [
    {
      "url": "/api",
      "description": "Local API server"
    }
  ],
  "paths": {
...
    "/community": {
      "post": {
        "summary": "Create a new community",
        "description": "Creates a new community record.",
        "operationId": "createCommunity",
        "tags": [
          "Community"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InsertCommunity"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Community created successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "data": {
                      "$ref": "#/components/schemas/Community"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "get": {
        "summary": "Get communities",
        "description": "Retrieves a list of communities based on provided filters.",
        "operationId": "getCommunities",
        "tags": [
          "Community"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "page",
            "schema": {
              "type": "integer",
              "default": 1
            },
            "description": "The page number for pagination."
          },
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "type": "integer",
              "default": 6
            },
            "description": "The number of items per page for pagination."
          },
          {
            "in": "query",
            "name": "activeTab",
            "schema": {
              "type": "string",
              "enum": [
                "all",
                "my"
              ],
              "default": "all"
            },
            "description": "Filter communities by tab (all or my)."
          },
          {
            "in": "query",
            "name": "createdByUserId",
            "schema": {
              "type": "string"
            },
            "description": "Filter communities by creator user ID."
          },
          {
            "in": "query",
            "name": "isPublished",
            "schema": {
              "type": "boolean"
            },
            "description": "Filter communities by published status."
          }
        ],
        "responses": {
          "200": {
            "description": "Communities retrieved successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "communities": {
                          "type": "array",
                          "items": {
                            "$ref": "#/components/schemas/Community"
                          }
                        },
                        "allCommunitiesPagination": {
                          "$ref": "#/components/schemas/Pagination"
                        },
                        "joinedCommunities": {
                          "type": "array",
                          "items": {
                            "$ref": "#/components/schemas/Community"
                          }
                        },
                        "joinedCommunitiesPagination": {
                          "$ref": "#/components/schemas/Pagination"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/community/slug-available/{slug}": {
      "get": {
        "summary": "Check if community slug is available",
        "description": "Checks if a given community slug is available for use.",
        "operationId": "isCommunitySlugAvailable",
        "tags": [
          "Community"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "slug",
            "schema": {
              "type": "string"
            },
            "required": true,
            "description": "The slug to check for availability."
          },
          {
            "in": "query",
            "name": "communityId",
            "schema": {
              "type": "string"
            },
            "required": false,
            "description": "Optional. The ID of the community to exclude from the check (for updates)."
          }
        ],
        "responses": {
          "200": {
            "description": "Slug availability check successful.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "data": {
                      "type": "boolean",
                      "description": "True if the slug is available, false otherwise.",
                      "example": true
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
  },
...
}